// src/TaTooIne.Demo/wwwroot/_ts/lib/lightweight-charts/dist/typings.d.ts
export declare enum CrosshairMode {
	Normal = 0,
	Magnet = 1
}
export declare enum LineStyle {
	Solid = 0,
	Dotted = 1,
	Dashed = 2,
	LargeDashed = 3,
	SparseDotted = 4
}
export declare enum LineType {
	Simple = 0,
	WithSteps = 1
}
export declare enum PriceLineSource {
	/**
	 * The last bar data
	 */
	LastBar = 0,
	/**
	 * The last visible bar in viewport
	 */
	LastVisible = 1
}
/**
 * Enum of possible price scale modes
 * Normal mode displays original price values
 * Logarithmic mode makes price scale show logarithms of series values instead of original values
 * Percentage turns the percentage mode on.
 * IndexedTo100 turns the "indexed to 100" mode on
 */
export declare enum PriceScaleMode {
	Normal = 0,
	Logarithmic = 1,
	Percentage = 2,
	IndexedTo100 = 3
}
export declare enum TickMarkType {
	Year = 0,
	Month = 1,
	DayOfMonth = 2,
	Time = 3,
	TimeWithSeconds = 4
}
// ...
export declare type AreaSeriesPartialOptions = SeriesPartialOptions<AreaStyleOptions>;
// ...
export declare type DeepPartial<T> = {
	[P in keyof T]?: T[P] extends (infer U)[] ? DeepPartial<U>[] : T[P] extends readonly (infer X)[] ? readonly DeepPartial<X>[] : DeepPartial<T[P]>;
};
// ...
export declare type OverlayPriceScaleOptions = Omit<PriceScaleOptions, 'visible' | 'autoScale'>;
// ...
export declare type Time = UTCTimestamp | BusinessDay | string;
export declare type TimeFormatterFn = (time: BusinessDay | UTCTimestamp) => string;
export declare type TimeRange = Range<Time>;
export declare type TimeRangeChangeEventHandler = (timeRange: TimeRange | null) => void;
export declare type UTCTimestamp = Nominal<number, 'UTCTimestamp'>;
export declare type VertAlign = 'top' | 'center' | 'bottom';
export declare type VisiblePriceScaleOptions = PriceScaleOptions;
// ...
export interface BarData {
	time: Time;
	open: number;
	high: number;
	low: number;
	close: number;
}
export interface BarPrices {
	open: BarPrice;
	high: BarPrice;
	low: BarPrice;
	close: BarPrice;
}
export interface BarStyleOptions {
	upColor: string;
	downColor: string;
	openVisible: boolean;
	thinBars: boolean;
}
export interface BusinessDay {
	year: number;
	month: number;
	day: number;
}
// ...
export interface ChartOptions {
	/** Width of the chart */
	width: number;
	/** Height of the chart */
	height: number;
	/** Structure with watermark options */
	watermark: WatermarkOptions;
	/** Structure with layout options */
	layout: LayoutOptions;
	/** Structure with price scale option for left price scale */
	leftPriceScale: VisiblePriceScaleOptions;
	/** Structure with price scale option for right price scale */
	rightPriceScale: VisiblePriceScaleOptions;
	/** Structure describing default price scale options for overlays */
	overlayPriceScales: OverlayPriceScaleOptions;
	/** Structure with time scale options */
	timeScale: TimeScaleOptions;
	/** Structure with crosshair options */
	crosshair: CrosshairOptions;
	/** Structure with grid options */
	grid: GridOptions;
	/** Structure with localization options */
	localization: LocalizationOptions;
	/** Structure that describes scrolling behavior or boolean flag that disables/enables all kinds of scrolls */
	handleScroll: HandleScrollOptions | boolean;
	/** Structure that describes scaling behavior or boolean flag that disables/enables all kinds of scales */
	handleScale: HandleScaleOptions | boolean;
}
// ...
export interface CrosshairOptions {
	/** Crosshair mode */
	mode: CrosshairMode;
	/** Options of the crosshair vertical line */
	vertLine: CrosshairLineOptions;
	/** Options of the crosshair horizontal line */
	horzLine: CrosshairLineOptions;
}
/** Structure describing horizontal or vertical grid line options */
export interface GridLineOptions {
	/** Color of the lines */
	color: string;
	/** Style of the lines */
	style: LineStyle;
	/** Visibility of the lines */
	visible: boolean;
}
/** Structure describing grid options */
export interface GridOptions {
	/** Vertical grid line options */
	vertLines: GridLineOptions;
	/** Horizontal grid line options */
	horzLines: GridLineOptions;
}
export interface HandleScaleOptions {
	mouseWheel: boolean;
	pinch: boolean;
	axisPressedMouseMove: AxisPressedMouseMoveOptions | boolean;
	axisDoubleClickReset: boolean;
}
export interface HandleScrollOptions {
	mouseWheel: boolean;
	pressedMouseMove: boolean;
	horzTouchDrag: boolean;
	vertTouchDrag: boolean;
}
/**
 * Structure describing a single item of data for histogram series
 */
export interface HistogramData extends LineData {
	/**
	 * Optional color value for certain data item. If missed, color from HistogramSeriesOptions is used
	 */
	color?: string;
}
export interface HistogramStyleOptions {
	color: string;
	base: number;
}
export interface IChartApi {
	/**
	 * Removes the chart object including all DOM elements. This is an irreversible operation, you cannot do anything with the chart after removing it.
	 */
	remove(): void;
	/**
	 * Sets fixed size of the chart. By default chart takes up 100% of its container
	 *
	 * @param width - target width of the chart
	 * @param height - target height of the chart
	 * @param forceRepaint - true to initiate resize immediately. One could need this to get screenshot immediately after resize
	 */
	resize(width: number, height: number, forceRepaint?: boolean): void;
	/**
	 * Creates an area series with specified parameters
	 *
	 * @param areaOptions - customization parameters of the series being created
	 * @returns an interface of the created series
	 */
	addAreaSeries(areaOptions?: AreaSeriesPartialOptions): ISeriesApi<'Area'>;
	/**
	 * Creates a bar series with specified parameters
	 *
	 * @param barOptions - customization parameters of the series being created
	 * @returns an interface of the created series
	 */
	addBarSeries(barOptions?: BarSeriesPartialOptions): ISeriesApi<'Bar'>;
	/**
	 * Creates a candlestick series with specified parameters
	 *
	 * @param candlestickOptions - customization parameters of the series being created
	 * @returns an interface of the created series
	 */
	addCandlestickSeries(candlestickOptions?: CandlestickSeriesPartialOptions): ISeriesApi<'Candlestick'>;
	/**
	 * Creates a histogram series with specified parameters
	 *
	 * @param histogramOptions - customization parameters of the series being created
	 * @returns an interface of the created series
	 */
	addHistogramSeries(histogramOptions?: HistogramSeriesPartialOptions): ISeriesApi<'Histogram'>;
	/**
	 * Creates a line series with specified parameters
	 *
	 * @param lineOptions - customization parameters of the series being created
	 * @returns an interface of the created series
	 */
	addLineSeries(lineOptions?: LineSeriesPartialOptions): ISeriesApi<'Line'>;
	/**
	 * Removes a series of any type. This is an irreversible operation, you cannot do anything with the series after removing it
	 */
	removeSeries(seriesApi: ISeriesApi<SeriesType>): void;
	subscribeClick(handler: MouseEventHandler): void;
	/**
	 * Removes mouse click subscription
	 *
	 * @param handler - previously subscribed handler
	 */
	unsubscribeClick(handler: MouseEventHandler): void;
	/**
	 * Adds a subscription to crosshair movement to receive notifications on crosshair movements
	 *
	 * @param handler - handler (function) to be called on crosshair move
	 */
	subscribeCrosshairMove(handler: MouseEventHandler): void;
	/**
	 * Removes a subscription on crosshair movement
	 *
	 * @param handler - previously subscribed handler
	 */
	unsubscribeCrosshairMove(handler: MouseEventHandler): void;
	/**
	 * Returns API to manipulate the price scale
	 *
	 * @param priceScaleId - id of scale to access to
	 * @returns target API
	 */
	priceScale(priceScaleId?: string): IPriceScaleApi;
	/**
	 * Returns API to manipulate the time scale
	 *
	 * @returns target API
	 */
	timeScale(): ITimeScaleApi;
	/**
	 * Applies new options to the chart
	 *
	 * @param options - any subset of chart options
	 */
	applyOptions(options: DeepPartial<ChartOptions>): void;
	/**
	 * Returns currently applied options
	 *
	 * @returns full set of currently applied options, including defaults
	 */
	options(): Readonly<ChartOptions>;
	/**
	 * Make a screenshot of the chart with all the elements excluding crosshair.
	 *
	 * @returns a canvas with the chart drawn on
	 */
	takeScreenshot(): HTMLCanvasElement;
}
// ...
export interface ISeriesApi<TSeriesType extends SeriesType> {
	/**
	 * Returns current price formatter
	 *
	 * @returns interface to the price formatter object that can be used to format prices in the same way as the chart does
	 */
	priceFormatter(): IPriceFormatter;
	/**
	 * Converts specified series price to pixel coordinate according to the series price scale
	 *
	 * @param price - input price to be converted
	 * @returns pixel coordinate of the price level on the chart
	 */
	priceToCoordinate(price: number): Coordinate | null;
	/**
	 * Converts specified coordinate to price value according to the series price scale
	 *
	 * @param coordinate - input coordinate to be converted
	 * @returns price value of the coordinate on the chart
	 */
	coordinateToPrice(coordinate: number): BarPrice | null;
	/**
	 * Retrieves information about the series' data within a given logical range.
	 *
	 * @param range - the logical range to retrieve info for
	 * @returns the bars info for the given logical range: fields `from` and `to` are
	 * `Logical` values for the first and last bar within the range, and `barsBefore` and
	 * `barsAfter` count the the available bars outside the given index range. If these
	 * values are negative, it means that the given range us not fully filled with bars
	 * on the given side, but bars are missing instead (would show up as a margin if the
	 * the given index range falls into the viewport).
	 */
	barsInLogicalRange(range: Range<number>): BarsInfo | null;
	/**
	 * Applies new options to the existing series
	 *
	 * @param options - any subset of options
	 */
	applyOptions(options: SeriesPartialOptionsMap[TSeriesType]): void;
	/**
	 * Returns currently applied options
	 *
	 * @returns full set of currently applied options, including defaults
	 */
	options(): Readonly<SeriesOptionsMap[TSeriesType]>;
	/**
	 * Returns interface of the pri scale the series is currently attached
	 *
	 * @returns IPriceScaleApi object to control the price scale
	 */
	priceScale(): IPriceScaleApi;
	/**
	 * Sets or replaces series data
	 *
	 * @param data - ordered (earlier time point goes first) array of data items. Old data is fully replaced with the new one.
	 */
	setData(data: SeriesDataItemTypeMap[TSeriesType][]): void;
	/**
	 * Adds or replaces a new bar
	 *
	 * @param bar - a single data item to be added. Time of the new item must be greater or equal to the latest existing time point.
	 * If the new item's time is equal to the last existing item's time, then the existing item is replaced with the new one.
	 */
	update(bar: SeriesDataItemTypeMap[TSeriesType]): void;
	/**
	 * Sets markers for the series
	 *
	 * @param data - array of series markers. This array should be sorted by time. Several markers with same time are allowed.
	 */
	setMarkers(data: SeriesMarker<Time>[]): void;
	/**
	 * Creates a new price line
	 *
	 * @param options - any subset of options
	 */
	createPriceLine(options: PriceLineOptions): IPriceLine;
	/**
	 * Removes an existing price line
	 *
	 * @param line - to remove
	 */
	removePriceLine(line: IPriceLine): void;
}
// ...
export interface LayoutOptions {
	/** Background color of the chart area and the scales */
	backgroundColor: string;
	/** Color of a text on the scales */
	textColor: string;
	/** Font size of a text on the scales in pixels  */
	fontSize: number;
	/** Font family of a text on the scales */
	fontFamily: string;
}
/**
 * Structure describing single data item for series of type Line or Area
 */
export interface LineData {
	time: Time;
	/**
	 * Price value of data item
	 */
	value: number;
}
export interface LineStyleOptions {
	color: string;
	lineStyle: LineStyle;
	lineWidth: LineWidth;
	lineType: LineType;
	crosshairMarkerVisible: boolean;
	crosshairMarkerRadius: number;
}
export interface LocalizationOptions {
	/**
	 * Current locale, which will be used for formatting dates.
	 * {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl#Locale_identification_and_negotiation}
	 */
	locale: string;
	/**
	 * User-defined function for price formatting. Could be used for some specific cases, that could not be covered with PriceFormat
	 */
	priceFormatter?: PriceFormatterFn;
	/**
	 * User-defined function for time formatting.
	 */
	timeFormatter?: TimeFormatterFn;
	/**
	 * Date formatting string.
	 * Might contains `yyyy`, `yy`, `MMMM`, `MMM`, `MM` and `dd` literals which will be replaced with corresponding date's value.
	 * Ignored if timeFormatter has been specified.
	 */
	dateFormat: string;
}
// ...
export interface TimeScaleOptions {
	rightOffset: number;
	barSpacing: number;
	fixLeftEdge: boolean;
	lockVisibleTimeRangeOnResize: boolean;
	rightBarStaysOnScroll: boolean;
	borderVisible: boolean;
	borderColor: string;
	visible: boolean;
	timeVisible: boolean;
	secondsVisible: boolean;
	tickMarkFormatter?: TickMarkFormatter;
}
/** Structure describing watermark options */
export interface WatermarkOptions {
	/** Color of the watermark */
	color: string;
	/** Visibility of the watermark. If false, other parameters are ignored */
	visible: boolean;
	/** Text of the watermark. Word wrapping is not supported */
	text: string;
	/** Font size in pixels */
	fontSize: number;
	/** Font family */
	fontFamily: string;
	/** Font style */
	fontStyle: string;
	/** Horizontal alignment of the watermark inside the chart area */
	horzAlign: HorzAlign;
	/** Vertical alignment of the watermark inside the chart area */
	vertAlign: VertAlign;
}